requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
ping3==5.1.5
psutil==7.2.2
speedtest-cli==2.1.3
Werkzeug==3.1.7
